Fixed some oversights
The game now actually checks if a card is currently in a pile before allowing you to add a card to a column from a pile (this would result in a nullptr error)

Fixed an issue with a character that was not supposed to be there appearing in the list of choices

Made it so the messages about player actions display after the console gets cleared instead of before so they actually are visible.

Added a 0 in front of the row numbers under 10 to maintain a coherent layout without alignement getting messed up when row 10 appears.

Added better text for victory.